+++ src/libpocketsphinx/ngram_search.h
@@ -306,7 +306,7 @@ struct ngram_search_s {
     int32 *word_lat_idx; /* BPTable index for any word in current frame;
                             cleared before each frame */
 
-    int32 *zeroPermTab; /**< Null right context table, just an array of
+    int16 *zeroPermTab; /**< Null right context table, just an array of
                            n_ciphone zeros (!!) */
 
     /*
